Don't dynamically allocate strings after all!
If an error occurs inside the OS call then allocated
memory isn't freed (it's like a C++ exception). This *could* be worked
around via setjmp() but then you'd need a big jmp_buf available, and it
would be just as easy to use that big buffer as a 256 byte workspace for
the temporary string. That may be the final solution, as this isn't
able to handle maximum length strings.