Fix longjmp() bug, tweak error handling
longjmp() didn't return the value to the caller of setjmp(); fixed, and
we also force 0->1 to ensure the setjmp() return value is unambiguous.

Use the longjmp() value to pass the error number to the caller

Don't store the result of top level error handler setjmp() in a byte;
this could misinterpret a 16-bit result with a low byte of 0 (although
in practice it's never a problem as all the code which longjmp()s to
this error handler passes an 8-bit error code as the value).