Refuse to relocate upwards
We could do this, but it would take more code and it seems unlikely to
be useful so let's wait and see if it ever causes a problem. We're
already loading so high that we overlap a mode 0-2 non-shadow screen, so
we "can't" just raise the initial load address from the current &2000.