Fix bug in relocation
We could cheerfully overwrite the download loop with random crap from
higher up in address space while it was executing. This wasn't obvious
on the Master but on an emulated B+ with DFS and ADFS and PAGE=&1F00, we
were only relocating down one page and it did cause problems. Quite
subtle problems - by sheer chance we seemed to end up in a loop which
push loads of copies of the flags onto the stack and then exited OK.
This meant that the setjmp() in execmod() was larger than our hacky 100
byte jmp_buf and the extra crap from the stack got copied onto the mode
7 screen, making the bug visible!