Try to walk up the stack as part of mode change
This "works" but it goes wrong; I am fairly sure this is because I am
assuming ENTER/LEAVE on each subroutine, whereas some will be -/RET and
those have one less item on the stack, and there's AFAICS no 100%
reliable way to determine which is which at this point. I want to commit
this change before I fiddle further.