Track the region of the FILE buffer which is dirty
My thinking here is that an application might have a "huge" buffer,
perhaps 10-20K or more, in an attempt to speed up random access to a
large file taking advantage of extra memory available. We don't want to
write out a 20K buffer in its entirety just because one byte changed
near the middle. Obviously there is still potential for poor behaviour
if you change a byte at the start and one at the end, but you can only
be so clever. To some extent an application needs to decide what its use
pattern looks like before using a very large buffer.