Use hash table in CS opcode string pool
This avoids the problem where using a CS opcode inside a loop causes the
same string to be repeatedly copied onto the frame stack. I haven't
benchmarked this but it probably doesn't cost very much; there's the
extra work to walk the linked list, but then the work of copying the
string is avoided. If you're doing this outside a loop it's probably not
a problem, and if you're doing this inside a loop you wouldn't get away
with it at all without this change.

The memory cost is 32 bytes per CS-using function for the hash table,
plus four extra bytes per CS opcode - all this is on the parameter stack
and is freed on function exit.