optimise.py: use tree optimiser first
Makes no difference at the moment but it may do later; gut feeling is
this order is better as it will give the tree optimiser a crack at
things before the peephole optimiser possibly confuses things. Hmm, not
so sure the more I think about it - but it's OK, at least.