Switch to new convention for OS calls
The 'raw' constants (e.g. osgbpb) are used for the OS entry point
addresses and the generally preferred approach is just to use the call()
function. A few special cases are still implemented in acornos and these
are disambiguated from the corresponding OS entry point by having a
call_ prefix on the name.

This could have broken something but I've run all the test programs and
they all seem to work, so fingers crossed it's all OK. I think this is
definitely something that needed doing at some point...