Fix bug in relocation code
If delta_relocation happened to be a multiple of 256, it could not be
encoded correctly. Say it was 256. We could skip along 256 bytes and
*not* perform a relocation using 0, and that's what we'd do, but we had
no way to encode "next relocation is in 256 bytes". So we need to use 0
to skip ahead 255 bytes; that works because we already have 255 to mean
"next relocation is 255 bytes ahead".

This caused intermittent failures; I noticed it while changing stocr to
assembler (next commit, though the actual code change isn't/wasn't at
fault, it's sort of an accident of how the code layout changes).