Experimentally turn optimiser on for PLAS32
No reason I'm aware of this couldn't be done for PLAS128 or any of the
other libraries/tests, but start small. This makes an extra 151 bytes of
space available; not huge, but not too shabby either.

I've tried all the test programs, including a game of ROGUE, and it
doesn't seem this has broken anything. I also reviewed the diff of the
optimiser output and it looks OK, though of course if it's emitting bad
hex opcodes I wouldn't notice, as I just read the comments with the
text opcodes.