Allow PLASMA to restart correctly on BREAK on second processor
I'm no expert here but this seems to work OK, including things like the
BRK handler after the reset. It's a minor shame to lose a few bytes of
workspace on the second processor, but it can't really be helped - it's
ugly if pressing BREAK leaves the machine in a weird state.