Fix PLAS128 bug in rdstr()
It was lowering IFP 5 bytes to get temporary workspace, but on PLAS128
this could trample on the calling function's sstring pool. Change to
using workspace at SCRATCH; this works on all builds and as a bonus is
shorter code as well. I noticed this bug in the code, it never actually
caused a problem.