Fix more call() ordering bugs
I did a review of all call() calls in the files I think are relevant to
the Acorn port, so fingers crossed this is the last of them. ROGUE now
works again, it didn't before. (I have tried twice and *instantly* die
on fighting a thief, but that may just be that the game is very hard...)