Use constant string in textscrn benchmark
This works fine in PLAS128 now we have a hash pool for CS, and it's a
"fairer" translation of the benchmark. Performance on PLAS128 (running
under an emulated Master in BeebEm) drops from 16.96s to 17.2s but
that's not too bad, and it's only the *benchmark* which has changed, not
the VM's performance.