Switch to upstream implementation of CS string pool
Much simpler than mine - the VM itself is 202 bytes smaller - and at
least on textscrn in slightly faster (was 17.2s, now 17.17s using
PLAS128 on BeebEm in Master mode). Also less memory used at runtime as
we don't have the 32 byte per-function-call hash table or the extra two
bytes for the next entry pointer on each string in the pool.

As a not inconsiderable bonus, this avoids the problem with mode() where
the pointers to strings in the hash table/linked lists in the pools on
the heap were not updated when the parameter stack was relocated,
because the string pool no longer contains any absolute addresses. (I
did have some ideas on how this could be worked around, but they would
have added further complexity.)

I suspect my string pool would be faster for code which is using many
strings or relatively long strings, but even if that's true this is
probably quite a rare and not performance critical case - and you can
always choose to use global strings if performance really does matter.