Make MACHID a byte in cmdsys.plh instead of a word
I think this is correct; most of the supplied Apple examples declare
MACHID themselves instead of importing this file, and they use byte not
word.

I think we got away with this before because the low byte of the word
was the correct value and the high byte was the low byte of cmdptr which
would usually be 0.