Teach optimiser about >> 8 optimisation for LAW and fix bug
The LLW code was working completely by chance due to no parentheses on
the SHR_CODE macro; I only noticed when this meant the new LAW code
happened to incorrectly perform a nonsensical optimisation on memmgr.a.
Have added missing brackets everywhere out of paranoia, this seems to
work.

The LAW optimisation is not actually trigger anywhere (but I did
temporarily add code to test it), but it costs nothing in the space
constrained VM and isn't a big complication in the PLASMA compiler so
why not?