Expose longer sequences to peephole optimiser
There are probably bugs but I've diffed the optimised output for a few
of the sample programs (including 32cmd.a, hanoi.a and sb.a) and there
are no non-comment differences, so this isn't working too badly.

By itself this does nothing, but I think this opens up the potential to
enable further optimisations.