Experimentally "un-mangle" some symbols
bbcmd.pla uses different internal names for some symbols compared with
the exported versions; this was inherited from the Apple code. I can't
see why this is necessary and as can be seen from the TODOs it is
problematic for my standalone build. Experimentally remove the
"mangling" for a couple of symbols; this doesn't appear to break
anything (standalone or module based) and even if the standalone
experiment is eventually abandoned I think this makes the code much more
readable.