Remove TODOs
Dave did the work for me to implement the self-modifying code
optimisation everywhere I'd identified it could be helpful. The 'ZEROL'
optimisation *is* a saving on the byte opcodes but not as good as
self-modifying code, and my gut feeling right now is that I should
follow upstream and not use it as the non-self-modifying alternative.
(While the VM is always going to be prone to breakage if code scribbles
over random memory, the idea of ZEROL accidentally getting corrupted to
non-0 and causing odd results scares me a bit too.) As noted by Dave,
the ZEROL idea doesn't help enormously with the word opcodes - I
experimented with this myself, it *may* be a small saving overall but
it's not great, the problem being that when you advance Y to access the
high byte of the operand, you have to check for it wrapping (INY:BNE
+:INC ZEROH:+) and that bloats the code and reduces the benefit. So
let's at least remove the TODOs about this and if I have any brainwaves
or feel a ROMmable version of the VM needs tweaking later on I can
always revisit this.