Tidy (over-engineer?) module name to filename translation
Not sure if this is OTT but let's go with it for now. The VM has 96
bytes less free at the prompt now, but part of this is due to the extra
error handling, not just the pluggable translation support - you can no
longer crash the VM by entering +THISISAVERYLONGSTRINGINDEEDANDNOMISTAKE
or similar at the prompt.