Handle ESCAPE in getc(), tidy labels
getc() no longer returns 27 every time if you press ESCAPE...

Labels are sort of less tidy but the '.' prefix for local symbols is a
bit misleading as an asm function doesn't automatically define a zone,
so they are not really local. The new names are fine and more consistent
with the rest of the PLASMA code.