Add setjmp2()
I've left some test programs using jbsz; I don't intend to get rid of
this, as it has no space cost at runtime when it isn't used and it is a
bit simpler to understand, which is useful for test/demo programs.

I suppose setjmp2() is a bit inefficient as its presence will mean there
are a few extra bytes on the CPU stack which need to be saved into the
jmp_buf, but I think it's probably worth it for its relative simplicity.