Fix over-eager code in remove_writes() (and rename it)
See the comments; I didn't ever actually run into these problems but I
believe they are completely possible. It's a shame as this will weaken
this optimisation even further; it may even be that it's doing next to
nothing now, but let's leave it in as long as we believe it's correct.

The renaming is unrelated, but it seemed inconsistent to use 'writes'
when we we use 'store' everywhere else.