Fix problems in standalone builds
I suspect divmod should have been exported before, but I got away with
it as I'd never tested a standalone program which included the
cmdsysac.plh header where it used to be declared - now it has moved into
cmdsys.plh it's more obvious that this was wrong.

The INITDSTSRCY move is probably another lingering bug; the common code
factored into the subroutine needs to be included in standalone builds,
so it has to live inside an asm function which isn't removed in those
builds.