Add JIT support to cmd.pla
Still all experimental hacks (I will stop saying that in each commit...)
cmd.pla is now capable of loading the JIT module on startup; for the
moment I am just using the HELLO module as a pseudo-JIT to demonstrate
that the VM loads and executes it correctly before showing the prompt.

The VM does still seem to work (I'm testing it by executing the
self-hosted compiler and running the resulting modules).