JIT partially enabled in Acorn VM
I haven't addressed the issue of the change in location of ESTK or the
different INTERP entry points, so I didn't expect this to work. However,
the VM starts and invoking the self-hosted compiler to build hello.pla
locks up roughly as I expected at the "SFTODO!" puts() in defcopy()
after working away for a little bit.