Debugging snapshot
With this exact build, setting a breakpoint on writes to 17E8 will show
the code just after the call to OSWORD in gets() being corrupted when
running "+jitune 1 255" then "+PLASM TEST.PLA". This explains why the
crash seems to happen *after* the compilation; the corruption caused
isn't hit until RETURN is pressed at the end of the following command.