pt.py snapshot
The output assembles and has a plausible size, but the self-hosted
compiler doesn't run - it *does* start up, but it misbehaves almost
immediately. I hope it will be obvious what the problem is tomorrow...

The code is pretty foul but if I can get this working I can then start
to try to come up with nicer data structures and try to split things up
into more logical classes etc.