pt.py snapshot
We now preserve the order of things in the input better. Doing this made
it obvious the primary problem was my hack to just use 999 for the value
of CW opcodes! However, I haven't tried the output yet as I think there
is a bug in my handling of case blocks; this is a snapshot before trying
to fix that.