pt.py now has load/store optimisation
Only lightly tested on self-hosted compiler but it does seem to work.
Not a huge saving (it removes five DLW opcodes) but the implementation
(if not buggy) is not too complex and it acts as some kind of validation
that my disassembled bytecode representation is reasonably convenient to
work with.