Experimental rewrite of optimise_load_store()
I haven't removed the old implementation yet. The output *doesn't* match
that of the previous code, but it seems to be much better - it's as if
the old code was buggy, in a way I haven't yet. Not sure if I will -
but it might be nice to figure it out in case I've overlooked something
in the new implementation. I also noticed what I think is a bug about
not treating LB/LW as capable of observing arbitrary stores, but that is
a bug carried over from the previous implementation and I have added a
TODO about it.

Going to sleep on this. Once I'm happy with it I can tidy it up further
(e.g. tidy up add_affect(), perhaps get it to return a set by value,
remove some of the special casing of 'simple' loads and stores) and try
to extend this to also operate on absolute loads/stores (LAW etc) as
well as frame loads/stores (LLW etc).