Optimise _INIT, support absolute addresses
We weren't optimising _INIT, d'oh! Self-hosted compiler still works now
we are doing that.

Added support for absolute addresses on MemoryInstructions (e.g. SAW
$3000). Have lightly tested this with a simple program (not the
self-hosted compiler, which doesn't use absolute addresses) and it
successfully optimises redundant SAW $3000 operations but not SAW $D000
operations (which are deemed non-redundant as that's considered an I/O
address).

Repurposed the previously dead Word class as the new Address class.

Code is a bit hacky but the whole thing is; this is still progress, I
think.