Rename Offset/"local label" to Target/target
Not sure this is a good idea, but it's at least more consistent and I
can always change it later. Offset wasn't a good name anyway, since the
branch targets are only expressed as offsets in the binary format.

A few other minor bits of tidying are mixed in with this.

Output is unchanged.