First cut at pt.py module splitting
pt.py is now capable of splitting off a single bytecode function, so we
have a PLASM3 module as well as the PLASM2 module it previously created.
The resulting PLASM2 module does seem to work, and by loading MODNAME
first I can see that it is pulling in the PLASM3 module.