pt.py splitting working tolerably well
There is no control over the target size of the two modules, but it just
happens to give a tolerable split on the PLASMA self-hosted compiler so
definitely worth a snapshot. The split compiler seems to work fine, both
on PLASMA (on second processor) and PLAS128 (on a Master 128). (The
data+code buffer size for the compiler on PLAS128 is disappointingly
small, but although there's probably nothing I can do about that, it is
enough to compile TEST.PLA at least.)

(A model B+ with DFS+ADFS and hence PAGE at &1F00 doesn't have enough
RAM free to load the compiler in PLAS128, though I'm guessing that there
would be a shortage of RAM for the data+code buffer anyway.)