Use shorter symbol names when splitting modules
This is useful because the symbol names remain in the symbol table and
reduce the main RAM available, so having them shorter reduces RAM use.
This works fine (as it did before) on a Master 128 under PLAS128. A
model B+ with PAGE at &1900 has 139 data+code bytes in the PLASMA
compiler, which is mostly useless, but the compiler is at least able to
load and run.