Remove TODO
I've checked the code and things only get into the ESD if they are
exported or if the dump of the RLD immediately before the ESD is dumped
puts them there. Essentially the non-exported part of the ESD is
generated from scratch after all optimisations are performed, so it
shouldn't include anything redundant.