Snapshot; working to get rid of callees() and use add_dependencies()
I had a funny feeling about having both as they seemed very similar yet
subtly different.

This is incomplete but works - it generates a "better" split of the
self-hosted compiler than the old code, which on investigation is caused
by the old code not allowing a function which calls itself (the
self-hosted compiler's nextln) to be moved into the split-off module.
I've tested the new split compiler on both PLASMA and PLAS128 and it
works.