Finish tidying split() code
This changes the output although it's exactly the same size as before; I
believe it's just the ESD which is differently ordered, so I've added
code to minimise this for future changes.

I've tested the new output (i.e. the self-hosted split compiler) on
PLASMA and PLAS128 and it seems to work fine.