Remove SFTODO
See scan() in C compiler sources.

I suppose arguably a module name might be able to start with a wider
range of characters but I don't think there's any way to comprehensively
establish that, and I'm not overly worried about this. (I don't believe
module names are prefixed with # in the symbol table any more, at least
not on the Apple and I will make the Acorn version match in due course.)