Improve optimiser control code
We weren't re-doing the step 1 optimisations before once step 2 had a
go; this meant we were missing some optimisation opportunities. The
output has changed but I've reviewed the differences and they all look
like improvements, and I've tested the self-hosted compiler binaries
still work.

I also experimentally merged all the optimisations into a single step
(not committed the code, of course) and as I thought this does generate
different code - the resulting code is still valid and is slightly
shorter, but it will be slightly slower because we replace a few
instances of "foo:branch y" with "branch x" where we have "target
x:foo:branch y". I am not sure the effects of having the optimisations
in one step or two steps will always be consistent, but this does at
least show it makes a difference and my feeling is I'll stick with what
I have.