Tidy up code formatting
I'm not super happy with this but it's probably better than it was. I've
tried to be guided by PEP-8 but my indentation of multi-line statements
is probably poor and inconsistent.

I've gone for a 90 character line length as a soft limit and allowed
myself to stray over that up to a maximum of 104 (my current and I think
typical terminal width) where it helped.