Optimise JITIINTERP and make it work correctly on PLAS128
By letting the IINTERP code handle the execution when we don't call the
JIT we automatically get the right behaviour on both PLASMA and PLAS128.
(This isn't to say the JIT works on PLAS128 yet; I need to make other
changes for that. But it's now possible to run the TEST sample on
PLAS128, whereas before it didn't work even though that never invokes
the JIT with the default JIT parameters.)

This is also shorter and faster than the old code; I don't think the
speed matters much since functions which aren't eligible for JITting or
which the JIT is unable to compile use IINTERP directly, but the speed
doesn't hurt and the saving on code size is nice.