Snapshot of experimental hack for PLAS128 JIT
This won't work (and I did test it just in case), I think because we
need code to copy bytecode from AUX into main RAM for the JIT to work
on, but while a bit of an unpleasant hack this approach may be basically
sound. I will at least try to get it working before considering
alternatives.