Interesting hack
I've hacked the JIT so it always refuses to compile anything. PLASM2 is
now able to compile TEST.PLA on PLAS128 without any problems. This
*suggests* (it may not be true) that the hacks I did to the core VM to
call the JIT are working, and that maybe the problem is either in the
JITted code itself or in the calling of the JITted machine code.