Use the superior and original Apple JMPTMP approach
I must have had a brain fart when I did the original port. I was trying
to conserve not-all-that-scarce zero page space but not having JMPTMP at
a fixed address has been an increasing pain, and I finally realised that
the Apple approach saves two cycles as well.

This does seem to work but I had terrible trouble as I kept forgetting
to update the various multiple constants in the source to reflect the
rerarranged zero page. I'm probably going to look at getting rid of
these duplicates next.