Rework page 4 allocations
Arguably a smidge over-complicated but I think it's fine, and I don't
like to waste space even if it's just a few extra bytes for the symbol
table. I haven't actually relocated the symbol table yet as I wanted to
get this working first and commit the code before going on.