Use all remaining space in pages 4, 5 and 6 for low symbol table chunk
Seems to work. It's probably time I stopped forcing the low symbol table
chunk to be small to test my assumption that I'm safe to grow the symbol
table as a module is loaded; it's been fine all this time, and I can
always revert this later if I want to experiment.

Updated SYMTBL to cope; this is actually nicer, as it now shows the
entire symbol table including the symbols built into the VM.