Factor out common code from ICAL and CALL
This will slow ICAL down by 3 cycles for the branch and the cycles for
CHECKEXPRESSIONSTACK, but we did the latter on CALL (which is way more
common) and only didn't do it on ICAL to save code size, so it seems
better this way.

This saves 39 bytes on PLAS128 and 22 bytes on PLASMA, which I think is
worth it. A few bytes of that on PLAS128 is due to the accidental code
duplication, but even so this seems a good deal.