Fix bug in PLAS128 CS opcode
The code assumed A was preserved by CHECKVSHEAPHIGHINA; it *usually* is
but not always. (I happened to trigger this on Master 128 by running
P128JIT in non-shadow mode 0 so the JIT can't load, then running
PLASM2.) A trivial reordering fixes this and I've checked all the other
uses of it and CHECKVSHEAP; I don't want to tweak the code to stack and
restore A as it's inside a macro and will bloat the code a bit.