Rewrite allocxheap()
This was a bit of a nightmare TBH but I think the previous version was
buggy and the new version is probably clearer and probably correct.
P128JIT now fails in a fairly clean way if you run it on a machine with
only 16K of sideways RAM, because it has exactly 0 bytes of AUX FREE
after loading the JIT and so running any other bytecode fails with a
heap full error.