Tweak bounce.pla
This makes it a bit nicer to look at and a bit more useful as a
benchmark. JIT seems to reduce the time taken for 256 frames by 44%
according to a quick play with b-em in Master 128 mode, which is nice.
(Pretty sure it still doesn't compete with COWGOL on performance
though...)