Show module name in "Not found" error
A nice usability improvement, well worth a bit of VM growth I think. I'm
using the module name rather than the full pathname because it's just so
much easier; the pathname is a CR string and could be very long, the
module name is a DCI string which we can easily translate using dcitos()
and can't be more than maxnamelen characters.