Remove TODO
This is a non-issue - it happens on the Apple II JIT VM because it loads
the JIT module high in AUX space and then needs to swizzle the address
down to start "user" code loading low. My approach of loading the JIT
module low and then knocking xheap up to the start of the next 16K bank
is roughly equivalent but doesn't require this swizzle at this point; I
wrote the TODO as part of my initial exploratory JIT work when I was
trying to get a lot of concepts straight in my head and didn't want to
accidentally shaft myself by failing to copy something critical from the
A2.