Alternative solution for command line parsing
This needs tidying but seems to work well, and it may even be smaller -
though I'd need to clean the code up and diff it with a
pre-experimental-hacks commit to see. Noticed while testing this that
doing "+ PLASM TEST.PLA" causes PLASM to crash in a nasty way but I
tried it on Apple II and the same thing happens there (admittedly it
kills the VM more on the Acorn, but I don't think it's that worrying;
once you start scribbling over memory it's all just pot luck), have
raised issue #53 upstream for this.

The basic idea here is that we leave the command line read from the user
alone - this means we haven't stripped the leading command character or
truncated it to just the module name - and so args.pla works without any
changes from the upstream code. Most things don't care about stripping
the leading command character (e.g. OSCLI quite happily accepts commands
which have a "*" in them) and we've reworked + so obviously it doesn't
care. We just copy the module name into a separate buffer; I've used
SCRATCH for this at the moment, but may review that (though it's
probably a sound choice with a small tweak or two).

"*PLASMA +PLASM TEST.PLA" to launch VM and run the compiler now works
without the append-a-CR hack I had in place at one point.