Shuffle some code around
This and probably the next few commits are preparing for implementing
the "docmds()" idea from the Apple II VM, which should allow a bit of
PLASMA init bytecode to be discarded on startup on non-tube VMs. (Tube
VM has to be able to restart on soft break, so it can't discard this
code I think, but we'll see if there's a few bytes we can get away with
- probably not though.) For the moment I am reshuffling code to get
ready but not changing the actual functionality or size of the code.