Basic docmds() for non-tube only; tube broken
Snapshot commit really. I thought this wasn't going to be worth it -
because we have to be able to handle OS errors as soon as we start
printing characters (we have to assume e.g. someone is using *SPOOL and
a disc error occurs, and we are the BRKV claimant so it's up to us) we
can't discard as much init code as we'd like, plus our symbol table
implementation means we don't copy it using one-off code as the Apple II
VM does. However, it seems to save a modest number of bytes which I find
hard to resist if I can make it work without too much effort.