Initialise heapptr directly in VM assembly code
This reduces the cost of docmds() on the tube (by getting rid of
lastdef() and simplifying the code at _INIT, neither of which we can
overwrite with the heap on the tube), is cleaner all round and has no
effect on the other configurations.

Tube version is now only 5 bytes worse off than it was before I started
this series of changes, though to be fair if I had a HIPLASMA it would
now be a few bytes better off than it started because the improvements
in this commit would be there without any of the cost of docmds().
However, the cost of docmds() is reduced by this change anyway.