Experiment with different split size for PLASM2/3
I had hoped that by keeping PLASM3 as small as possible the extremely
long "link time" after loading the modules from disk would be minimised.
I haven't done any precise timings and it may be that using the
commented-out 0x4b00 value *is* an improvement, but it's very long and
for all I know longer. I seem to have lost an SFTODO before committing
which observed that the size() returned is approximate, because we
haven't done the export fixup after the while loop when we're calling
size(). The 0x4b00 value is experimentally (going down 0x100 at a time)
the largest a Master 128 can load under PLAS128 in shadow screen modes.
This wouldn't work on a B or B+ even in shadow modes, whereas PLASM2
"fully split" does work on those (not tested B, but have tested B+ in
shadow mode - it can't compile test.pla, but there's enough RAM to be
moderately useful).

Anyway, I'm not writing this off but since the improvement is not
astounding I'm reluctant to change the nature of the split between
PLASM2/3 just now.