Snapshot of library support
This is a mess but it seems to have some promise; I want to commit this
before playing around more.

Bloating the core VM by exporting lookupsym() is a bit annoying, but I
think it's worth it for being able to write resident modules which are
able to behave as lib.pla is doing any not re-install themselves over
and over again. It would be possible to copy the trick used in
symtab.pla to avoid the need for this, but that feels very hacky now
it's being done in "real" not-just-demo code and I wouldn't want to
encourage hypothetical users to embed knowledge of how the symbol table
is stored in their code by copying symtab.pla's method.

This is also less bloat than implementing the library support directly
in the core VM where it has to be paid for even if it isn't used.