Optimise memxcpy...() in PLAS128 JIT
This will be microscopically slower and eats two bytes of page 4 (only
on the PLAS128 JIT build, so nothing else pays this price) but it
shrinks the assembly blob which will take up valuable main RAM, so
definitely a worthwhile trade-off.

As per mildly cryptic comment, if I were willing to rely on my
experimental hack to allow "word foo" to generate fixups in asm
functions (as currently used acorn/fileio.pla), we wouldn't even need
those two bytes of page 4 (and we'd save a couple of totally irrelevant
cycles by using a direct instead of indirect JMP).