Add SFTODO
The bug in "+CAT P.TESTLIB" was present on the jit2 branch, so it's not
new. I still haven't figured it out, although adding some debug code in
fillbuf() to OSWRCH each byte we put into OUTBUF shows that we seem to
be inserting the correct bytes there, they are somehow being corrupted
before we get to puts().