Fix bug in cat.pla
This was actually using the *value* of the byte at 'buf' where it should
have been using the address of 'buf'. This meants that in practice the
buffer was at $0000, i.e. in OS zero page, and we mostly seemed to get
away with it, except that the P.TESTLIB case would show it up.

(To be honest, I'm not super confident about the way PLASMA-the-language
works here, but looking at the bytecode generated by the compiler it's
clear what was happening, and I originally discovered this by debugging
the fileio code using b-em's debugger and seeing that OUTBUF was at
$00xx. fileio.pla itself is not at fault here.)