Get rid of dword library and replace with int32
Mostly. I haven't completely purged dword yet, but I plan to at some
point. The int32 library is pretty nice, a lot more comprehensive and
of course it's good not to be gratuitously incompatible with upstream.

intmath32 benchmark is significantly improved by this, although the
benchmarks in general are much better than when I posted "plasma3" to
stardot at end of 2016 and I haven't double checked the improvement
isn't due to other changes. I'm pretty sure int32 library is better
optimised than dword was (I can see for a start that it has an unrolled
loop for addition unlike dword).