Tweak CS opcode, don't use manual string pool in cmd.pla
This seems to work and gives more RAM free at the prompt in both PLASMA
and PLAS128. Not a lot, but every little helps, and I think it's also
clearer in cmd.pla anyway. The CS opcode will be a few cycles slower in
PLAS128 but it's not performance critical.

This should also reduce minor weirdness in PLAS128 where the free memory
varies a little bit at the prompt depending on the sequence of OS errors
and strings printed altering the memory taken by the strig pool.