Add fileio version of store benchmark
This performs best of all three, which is quite reassuring. In hindsight
this is obvious, as it's the only one which (because of the work done by
the fileio library) will perform OS calls to write in 256 byte blocks.

Apart from being vaguely cool, this might also be helpful in
benchmarking any changes to fileio code, which I really need to tidy up
a bit, particularly with respect to error handling.