Tidy Acorn fileio.pla a bit more
Lightly tested (BENCH2 and PLASM)

I'm trying to use the ENHANCEMENT "tag" for things I might in theory do
in the future but don't intend to do in the short term, reserving SFTODO
for shorter term tweaks. (I'm not sure a "full" ProDOS emulation is
appropriate in here, given the differences between it and the various
Acorn filesystems and the lack of an portable application which would
use all the facilities in here if I implemented them. If one turns up
that's a different matter, of course, but as it is at the moment I'm
mainly interested in supporting enough for basic use like BENCH2 and for
the self-hosted PLASM compiler.)