Tweak Acorn fileio.pla write()
This is a few bytes shorter and I think the code is slightly clearer
because it avoids having a separate "name" for iobuf=>s_addr. It will be
a few cycles slower if we go round the loop more than once, but I really
can't see that being measurable nor have I been that paranoid about the
odd cycle in the rest of the code.

Lightly tested.