Add SFTODO/temporary note
Flirting with adding a double-buffered screen for less flickery updates,
so want to see how tight space is. If I haven't got confused (I didn't
write any notes) this is the tighest case that works - next is the same
config but in shadow mode 6 (or 7, but there's no point) which has an
extra 1K free of course, then PLAS128+PAGE=&1900+mode 7 or shadow mode,
which has about 5000 bytes free.